% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_pptx.R
\name{imp_pptx}
\alias{imp_pptx}
\title{导入Microsoft PowerPoint PPTX}
\usage{
imp_pptx(pptx_file, author = "hulilei")
}
\arguments{
\item{pptx_file}{pptx完整的文件路径}

\item{author}{作者，默认为hulilei}
}
\value{
返回值
}
\description{
导入Microsoft PowerPoint PPTX
}
\examples{
imp_pptx('aa.pptx','hulilei');
}
