% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{importPkgs}
\alias{importPkgs}
\title{导入包}
\usage{
importPkgs(conn = tsda::conn_rds("rdbe"), FappId = "system")
}
\arguments{
\item{conn}{连接}

\item{FappId}{软件程序}
}
\value{

}
\description{
导入包
}
\examples{
importPkgs()
}
