% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioConductor.R
\name{install_bioconductor}
\alias{install_bioconductor}
\title{构建相应的安装生成器,bioconductor}
\usage{
install_bioconductor(pkgName)
}
\arguments{
\item{pkgName}{名称}
}
\value{
返回值
}
\description{
构建相应的安装生成器,bioconductor
}
\examples{
#
}
