% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_china_tongji}
\alias{install_china_tongji}
\title{install from tongji}
\usage{
install_china_tongji(pkg)
}
\arguments{
\item{pkg}{pkgname}
}
\value{
return value
}
\description{
install from tongji
}
\examples{
install_china_tongji()
}
