% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlDms.R
\name{mdl_create}
\alias{mdl_create}
\title{创建DMS模块}
\usage{
mdl_create(
  mdl_name = "mdlStarNight",
  mdl_dir = "/home/hulilei/mdls/",
  demo_dir = "/home/hulilei/mdls/mdlDemo/",
  demo_name = "mdlDemo",
  type = c("Com", "Pkg", "r", "Server", "UI")
)
}
\arguments{
\item{mdl_name}{模块名称}

\item{mdl_dir}{模块目录}

\item{demo_dir}{模板目录}

\item{demo_name}{模板名称}

\item{type}{模板类型}
}
\value{
返回值
}
\description{
创建DMS模块
}
\examples{
mdl_create()
}
