% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlDms.R
\name{mdlCom_tidy}
\alias{mdlCom_tidy}
\title{修改com}
\usage{
mdlCom_tidy(mdl_dir = "/home/hulilei/mdls/", mdl_name = "mdlStarNight")
}
\arguments{
\item{mdl_dir}{修改路径名}

\item{mdl_name}{名称}
}
\value{
返回值
}
\description{
修改com
}
\examples{
mdl_tidy
}
