% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlDms.R
\name{mdlUI_tidy}
\alias{mdlUI_tidy}
\title{修改UI}
\usage{
mdlUI_tidy(
  mdl_dir = "/home/hulilei/mdls/",
  mdl_name = "mdlStarNight",
  version = "1.0.0",
  author = "hulilei"
)
}
\arguments{
\item{mdl_dir}{目标}

\item{mdl_name}{名称}

\item{version}{版本}

\item{author}{作者}
}
\value{
返回值
}
\description{
修改UI
}
\examples{
mdlUI_tidy
}
