% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_push2Github}
\alias{pkg_push2Github}
\title{代码推荐至github}
\usage{
pkg_push2Github(github_userName = "takewiki", branch = "main", init = TRUE)
}
\arguments{
\item{github_userName}{github用户名}

\item{branch}{分支}

\item{init}{初始化}
}
\value{
返回值
}
\description{
代码推荐至github
}
\examples{
pkg_push2Github()
}
