% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_push2Github_master_update}
\alias{pkg_push2Github_master_update}
\title{代码推荐至github的master分支}
\usage{
pkg_push2Github_master_update()
}
\value{
返回值
}
\description{
代码推荐至github的master分支
}
\examples{
pkg_push2Github()
}
