% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regAddin.R
\name{regAddin}
\alias{regAddin}
\title{注册Rstudio插件文件}
\usage{
regAddin(
  addinName = "注册Rstudio插件",
  addinDescription = "注册Rstudio插件，写入inst-rstudio-addins.dcf文件.",
  addBinding = "regAddin",
  addinInteractive = "false"
)
}
\arguments{
\item{addinName}{插件名称}

\item{addinDescription}{插件描述}

\item{addBinding}{插件函数名}

\item{addinInteractive}{插件是否交互模式}
}
\value{
返回值
}
\description{
注册Rstudio插件文件
}
\examples{
regAddin();
}
